requirements.txt
mmh3
requests
shodan
beautifulsoup4
python-dotenv
python-whois